Refactor and use cmake

+++ src/stream.cpp
@@ -1,5 +1,5 @@
-#include "includes/stream.hpp"
-#include "includes/quantization.hpp"
+#include "../inc/stream.hpp"
+#include "../inc/quantization.hpp"
 
 template<typename T>
 Stream<T>::Stream(unsigned long size):buffer(size) {
@@ -25,7 +25,7 @@ void Stream<T>::write(T d) {
 
 template<typename T>
 void Stream<T>::print() {
-  for(int i=0; i<size; i++) {
+  for(unsigned long i=0; i<size; i++) {
     cout << (int)buffer[i] << endl;
   }
 }
